requirements-dev.txt
# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.28.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
